Made minor changes to keep the now-pedantic compiler from griping. Also optimized the code a bit (took out unneeded .h includes, changed a few integers to special uint8_t etc.)

+++ ChaprSVN/Firmware/ChapR/settings.h
@@ -95,8 +95,8 @@ public:
      short analog4;
      bool matchModeEnable;
 
-     void doSetting(int,const __FlashStringHelper *, const __FlashStringHelper *,unsigned int,unsigned int,int);
-     void printCurrentValue(int,unsigned int, int);
+     void doSetting(int,const __FlashStringHelper *, const __FlashStringHelper *,unsigned int,unsigned int,uint8_t);
+     void printCurrentValue(int,unsigned int, uint8_t);
      int  getStringFromMonitor(char*, int);
      void setString(int, int, char*);
      char *getString(int, int);
